# Leverage C# 13 to simplify and optimize async serialization