--- a/src/Nerdbank.MessagePack/MessagePackAsyncWriter.cs
+++ b/src/Nerdbank.MessagePack/MessagePackAsyncWriter.cs
@@ -153,4 +153,15 @@ public class MessagePackAsyncWriter(PipeWriter pipeWriter)
 	{
 		return pipeWriter.CanGetUnflushedBytes && pipeWriter.UnflushedBytes > context.UnflushedBytesThreshold;
 	}
+
+	/// <summary>
+	/// Gets a value indicating whether it is time to flush the pipe.
+	/// </summary>
+	/// <param name="context">The serialization context.</param>
+	/// <param name="syncWriter">The synchronous writer that may have unflushed bytes to consider as well.</param>
+	/// <returns><see langword="true" /> if the pipe buffers are reaching their preferred capacity; <see langword="false" /> otherwise.</returns>
+	public bool IsTimeToFlush(SerializationContext context, in MessagePackWriter syncWriter)
+	{
+		return pipeWriter.CanGetUnflushedBytes && (pipeWriter.UnflushedBytes + syncWriter.UnflushedBytes) > context.UnflushedBytesThreshold;
+	}
 }