… src/Nerdbank.MessagePack/MessagePackAsyncReader.cs
… src/Nerdbank.MessagePack/PublicAPI.Unshipped.txt
… src/Nerdbank.MessagePack/StandardVisitor.cs
… test/Nerdbank.MessagePack.Tests/RawMessagePackTests.cs