@AArnott 2024-11-13
Code coverage for the async paths isn't where it should be by the time we stabilize. But I'm holding out writing tests to specifically cover these lines until I'm sure those lines will stick around.